Restructure config files to include providers
