+++ fitness-aggregator-api/src/main/java/com/jasonghent98/fitness_aggregator_api/config/provider/strava/StravaInterceptorConfig.java
@@ -1,0 +1,4 @@
+package com.jasonghent98.fitness_aggregator_api.config.provider.strava;
+
+public class StravaInterceptorConfig {
+}
